Add files via upload
This is a project I did using advanced excel techniques. The project is about to perform data analysis on a liquor store.

The data is organized in a way that follows a story. This is a business owner who wants to open a liquor store and used financial data to determine what his best investment methods would be. Once the store is purchased, he now has a list of products to purchase. Once the store has made some sales and revenue, there needs to be an analysis performed to view how the store is performing. The store eventually expands to two other locations and analyzes those store's performances. A few years later, due to pandemics, the store is forced to undergo a budget cut for their employee's payroll. The business owner also creates a rewards membership program whereby a user fills out a form and collects customer data. A worksheet collects customer reviews and has a gauge chart to visualize how the business' customers like the store's services.